% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditor_list.R
\name{auditor_list}
\alias{auditor_list}
\title{List audit records}
\usage{
auditor_list(x, connection = NULL)
}
\arguments{
\item{x}{List representing a filter for records}

\item{connection}{DBI database connection}
}
\value{
A list of records
}
\description{
List audit records
}
