% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditor_record.R
\name{auditor_record}
\alias{auditor_record}
\title{Get audit record}
\usage{
auditor_record(x, connection = NULL)
}
\arguments{
\item{x}{Record identifier}

\item{connection}{DBI connection}
}
\value{
List containing the audit record
}
\description{
The returned list of named elements consists of three main parts. The main
audit record, the record attributes and lastly links to other audit records.

The named element \code{id} refers to the audit record identifier. A record
attribute does not have an associated attribute identifier, but is referred
to by its key. Key values are not unique within the list of attributes and
serves only as a programmatic identifier.

Attributes are returned in natural sort order of the value for \code{key}.
}
