% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditor_save.R
\name{auditor_save}
\alias{auditor_save}
\title{Utility function to save one or more records}
\usage{
auditor_save(x, connection = NULL)
}
\arguments{
\item{x}{List of records}

\item{connection}{DBI database connection}
}
\value{
Logical if the records were created
}
\description{
This utility function will save an audit record along with its attributes.

A record is defined by a set of standard notations.
\itemize{
\item \code{event} the audit event
\item \code{type} the type of audited object
\item \code{class} optional reference to a class of object type
\item \code{reference} reference to the object
\item \code{object} optional hash representing the object regardless of context
\item \code{label} short human readable description of the event
\item \code{actor} user or service that initiated/triggered the event
\item \code{attributes} a list of attributes associated with the event
}

The \code{event} is one of the case in-sensitive values \code{create}, \code{read}, \code{update},
\code{delete}, \code{execute}, \code{sign}, \code{connect} or \code{disconnect}.

If \code{class} is not specified for an audited event, the value of \code{type} is used.

If \code{object} is omitted, the object hash is derived based on the specified \code{type},
\code{class} and \code{reference}.

\code{attributes} is a list of qualified key/value pairs. An attribute is defined
as
\itemize{
\item \code{key} attribute name represented as a keyword
\item \code{label} short human readable label of the key
\item \code{qual} or \code{qualifier} that provide context to the value
\item \code{value} the attribute value
}

The \code{label} and \code{value} values are stored URL encoded.

If the audit event record property, say the audit record \code{label}, is longer
than the property storage length, the property value is truncated in the main
record and the original value recorded as a record attribute with the \code{key}
equal to the propery name, the attribute \code{label} equal to
\verb{<property> original value}, \code{qualifier} equal to \code{original_value} and
attribute \code{value} equal to the original property value.

If more than one record is submitted, the individual audit records are linked.
}
