% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_appnode.R
\name{dev_appnode}
\alias{dev_appnode}
\title{(Development) Identify node by unique R session ID}
\usage{
dev_appnode()
}
\value{
Vector of identifying attributes
}
\description{
(Development) Identify node by unique R session ID
}
