% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\title{Start auditor service}
\usage{
start(port = 12345)
}
\arguments{
\item{port}{Service port}
}
\description{
A standard configurable start for plumber.

The function searches for \code{plumber.R} in the following directories
\itemize{
\item current working directory
\item directory \code{inst/ws} in the current working directory (development mode)
\item directory \code{ws} in the current working directory (installed)
\item directory \code{ws} in the first occurrence of \code{auditor.service} in the library tree
}
ws directory starting with
the current working directory and then package install locations in
\code{.libPaths()}.

The following app properties are used for database connections.
\itemize{
\item \code{DB.VENDOR} (future) to support vendor specific SQL syntax and optimizations
\item \code{DB.DRIVER} Database driver (DBI compatible)
\item \code{DB.DATABASE} Database name
\item \code{DB.HOST} Database host (defaults to \code{localhost})
\item \code{DB.PORT} Database port
\item \code{DB.USERNAME} Database username or account
\item \code{DB.PASSWORD} Database username or account password
\item \code{DB.POOL.MINSIZE} Minimum connections in database pool
\item \code{DB.POOL.MAXSIZE} Maximum connections in database pool
\item \code{DB.POOL.IDLETIMEOUT} Idle duration in seconds until dropping idle connections
beyond the pool's minimum number of connections
}

Note that the above configuration uses \link[cxapp]{cxapp_config} and supports
both environmental variables and key vaults.

For further details on \code{DB.POOL.MINSIZE}, \code{DB.POOL.MAXSIZE} and \code{DB.POOL.IDLETIMEOUT},
see \link[pool]{poolCreate}.
}
